feat: split age and gender pages
