+++ pages/quiz/age.tsx
@@ -6,12 +6,10 @@ import { useRouter } from 'next/router'
 import { Layout } from '../../src/Components/Layout'
 import { HeaderComponent } from '../../src/Components/Header'
 import { VerticalSpacing } from '../../src/Components/VerticalSpacing'
-import { Checkbox } from '../../src/Components/Checkbox'
 import { RadioButton } from '../../src/Components/RadioButton'
 import { QuizContext } from '../../src/context/QuizContext'
 import { LinkButton } from '../../src/Components/LinkButton'
 
-import { Gender } from '../../cms/services'
 import { StickyNavBar } from '../../src/Components/StickyNavBar'
 
 const Navigation = styled.section`
@@ -32,28 +30,6 @@ const QuestionSection = styled.section`
     border-bottom: 1px solid ${(props) => props.theme.colours.yellow};
     font-family: 'Catamaran', sans-serif;
     font-size: ${(props) => props.theme.fontSizes.heading};
-  }
-
-  fieldset {
-    border: none;
-    padding: 0;
-  }
-`
-
-const CheckboxGroup = styled.div`
-  max-width: 50ch;
-  margin: auto;
-  padding: 1rem var(--gutter-width);
-  width: 100%;
-
-  legend {
-    border: 0;
-    clip: rect(1px, 1px, 1px, 1px);
-    height: 1px;
-    overflow: hidden;
-    padding: 0;
-    position: absolute;
-    width: 1px;
   }
 
   fieldset {
@@ -93,40 +69,16 @@ const StyledLink = styled.a`
   }
 `
 
-const HorizontalGroup = styled.div`
+const RadioGroup = styled.div`
   margin: 1rem 0;
-  display: flex;
-  justify-content: space-between;
 `
 
-const categories: Array<Gender> = [
-  'men',
-  'women',
-  'non-binary',
-  'transgender',
-  'intersex',
-  'gender non-conforming',
-  'genderqueer',
-  'agender',
-]
-
-const allAges = ['<15', '15', '16', '17', '18', '18+']
+const allAges = ['under 15', '15', '16', '17', '18', 'over 18']
 
 export const AboutYouPage = (): JSX.Element => {
   const router = useRouter()
 
-  const {
-    ageGet,
-    ageSet,
-    genderGet,
-    genderToggle,
-    setQuizComplete,
-  } = useContext(QuizContext)
-
-  const skipQuestionAndSeeResults = () => {
-    setQuizComplete(true)
-    router.push('results')
-  }
+  const { ageGet, ageSet } = useContext(QuizContext)
 
   return (
     <Layout>
@@ -145,10 +97,10 @@ export const AboutYouPage = (): JSX.Element => {
           <LinkButton
             textContent="skip this question"
             arrow="right"
-            onClick={() => skipQuestionAndSeeResults()}
+            onClick={() => router.push('gender')}
           />
           <VerticalSpacing size={1} />
-          <HorizontalGroup>
+          <RadioGroup>
             {allAges.map((item) => {
               return (
                 <RadioButton
@@ -160,34 +112,12 @@ export const AboutYouPage = (): JSX.Element => {
                 />
               )
             })}
-          </HorizontalGroup>
+          </RadioGroup>
         </fieldset>
       </QuestionSection>
 
-      <CheckboxGroup>
-        <p>
-          Some support and activities in your area are gender specific. To find
-          something right for you, are you interested in services that are:
-        </p>
-        <VerticalSpacing />
-        <fieldset>
-          <legend>Are you interested in services for:</legend>
-          {categories.map((category) => {
-            return (
-              <Checkbox
-                key={category}
-                label={category}
-                checked={genderGet(category)}
-                onChange={() => genderToggle(category)}
-              />
-            )
-          })}
-        </fieldset>
-      </CheckboxGroup>
-      <VerticalSpacing />
-
       <Link href="/quiz/results" passHref>
-        <StyledLink onClick={() => setQuizComplete(true)}>{'Ok'}</StyledLink>
+        <StyledLink onClick={() => router.push('gender')}>{'Ok'}</StyledLink>
       </Link>
 
       <VerticalSpacing />
